+++ src/main/commands/scan_async.cc
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2013-2018 Aerospike, Inc.
+ * Copyright 2013-2020 Aerospike, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -76,5 +76,6 @@ NAN_METHOD(AerospikeClient::ScanAsync)
 
 Cleanup:
 	delete cmd;
+	if (p_policy && policy.base.predexp) as_predexp_list_destroy(policy.base.predexp);
 	as_scan_destroy(&scan);
 }