+++ src/main/commands/operate_async.cc
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2013-2018 Aerospike, Inc.
+ * Copyright 2013-2020 Aerospike, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -87,4 +87,5 @@ Cleanup:
 	delete cmd;
 	if (key_initalized) as_key_destroy(&key);
 	if (operations_initalized) as_operations_destroy(&operations);
+	if (p_policy && policy.base.predexp) as_predexp_list_destroy(policy.base.predexp);
 }